# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Freelancers",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://api-freelancer.otterscompany.com"
        }
    ],
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get categories",
                "description": "Returns all categories",
                "operationId": "indexCategory",
                "responses": {
                    "200": {
                        "description": "Show all categories.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Register category",
                "description": "Create a new category",
                "operationId": "storeCategory",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Category parent ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registered category.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    },
                    "422": {
                        "description": "Data not valid."
                    },
                    "500": {
                        "description": "An error occurred during the process."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category",
                "description": "Returns a single category",
                "operationId": "showCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search for ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show category.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    },
                    "404": {
                        "description": "No results found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update category",
                "description": "Update a single category",
                "operationId": "updateCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search for ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Category parent ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registered category.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    },
                    "422": {
                        "description": "Data not valid."
                    },
                    "500": {
                        "description": "An error occurred during the process."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete category",
                "description": "Delete a single category",
                "operationId": "destroyCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search for ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete category.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    },
                    "404": {
                        "description": "No results found."
                    },
                    "500": {
                        "description": "An error occurred during the process."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/categories/{id}/deactivate": {
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Deactivate category",
                "description": "Deactivate a single category",
                "operationId": "deactivateCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search for ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deactivate category.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    },
                    "404": {
                        "description": "No results found."
                    },
                    "500": {
                        "description": "An error occurred during the process."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/categories/{id}/activate": {
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Activate category",
                "description": "Activate a single category",
                "operationId": "activateCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search for ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activate category.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "403": {
                        "description": "Forbidden."
                    },
                    "404": {
                        "description": "No results found."
                    },
                    "500": {
                        "description": "An error occurred during the process."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}